requirements.txt
hyperliquid-python-sdk
eth-account
pandas
numpy
requests
schedule
python-dotenv
termcolor
ccxt
